requirements.txt
dj-database-url==0.5.0
Django==2.2.7
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
whitenoise==4.1.4
asgiref==3.2.3
django-ckeditor==5.8.0
django-crispy-forms==1.8.1
django-extensions==2.2.5
django-js-asset==1.2.2
six==1.13.0
sqlparse==0.3.0
